Fix build errors using gcc 4.6 with -std=c++0x.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10321 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/grapemcrl2/libgrape/coordinate.h
@@ -24,8 +24,8 @@ namespace libgrape
  */
 struct coordinate
 {
-  float m_x; /**< x-coordinate. */
-  float m_y; /**< y-coordinate. */
+  double m_x; /**< x-coordinate. */
+  double m_y; /**< y-coordinate. */
 
   /** @return True if the given coordinate is the same as the object. */
   bool operator==(const coordinate& p_c);
